fixed bug 358678
This commit contains the fixed ATL.acg, as well as the corresponding
ATLToASMCompiler.asm in plugin org.eclipse.m2m.atl.engine, which is used
for the compilation of .atl files. The copies of this file (ATL.asm
alongside ATLToASMCompiler.asm, and in plugin org.eclipse.m2m.atl.dsls)
that are apparently not in active use have not been updated.

This commit also contains the new compiled version of the
CreateManyTraceElements test case, which now passes.