fixed reverse bindings issues, notably bug 411681
The issues fixed in this commit are:
- bug 411681 about no resolution taking place in non-refining mode
- reverse bindings not applied at the right time: if reverse bindings
are executed in matching phase, elements created by rules matched after
(i.e., located after in source file) cannot be resolved. This is less of 
a problem in refining mode, although it could happen (e.g., for newly 
created elements that are resolved explicitly (i.e., using resolveTemp).
- reverse bindings with multivalued source not properly executed
- reverse bindings not inherited